+++ eniric/__init__.py
@@ -3,20 +3,26 @@ __version__ = "1.0rc1"
 import os
 import warnings
 
+eniric_dir = os.path.dirname(__file__)
+DEFAULT_CONFIG_FILE = os.path.join(eniric_dir, "config.yaml")
+
 from ._config import Config
 
-if os.path.exists("config.yaml"):
+if (
+    os.path.exists("config.yaml")
+    and os.path.abspath("config.yaml") != DEFAULT_CONFIG_FILE
+):
     config = Config("config.yaml")
 else:
-    base_dir = os.path.dirname(__file__)
-    default = os.path.join(base_dir, "config.yaml")
     warnings.warn(
-        "Using the default config.yaml file located at {0}. "
-        "This is likely NOT what you want. Please create a similar "
-        "'config.yaml' file in your current working directory.".format(default),
+        "Using the default config.yaml file located at {0}. This is likely NOT what you want and "
+        "you will not be able to change any of the config values. Please use config.copy_file(<path>) to copy a "
+        "version of the default config for your own project.".format(
+            DEFAULT_CONFIG_FILE
+        ),
         UserWarning,
     )
-    config = Config(default)
+    config = Config(DEFAULT_CONFIG_FILE)
 
 __all__ = [
     "atmosphere",