+++ eniric/broaden.py
@@ -18,10 +18,11 @@ from numpy.core.multiarray import ndarray
 from tqdm import tqdm
 
 import eniric
+from eniric import config
 from eniric.utilities import band_selector, mask_between, wav_selector
 
 # Cache convolution results.
-memory = Memory(location=eniric.config.cache["location"], verbose=0)
+memory = Memory(location=config.cache["location"], verbose=0)
 
 c_kmps = const.c.to("km/s").value
 num_procs_minus_1 = os.cpu_count() - 1