+++ eniric/utilities.py
@@ -2,13 +2,14 @@
 import collections
 import errno
 import os
+from os.path import join
 from typing import Any, List, Optional, Sequence, Tuple, Union
 
 import numpy as np
 from astropy import constants as const
 from numpy import ndarray
 
-import eniric
+from eniric import config
 
 # Band limits.
 bands_ = {
@@ -23,7 +24,7 @@ bands_ = {
     "NIR": (0.83, 2.35),
 }
 
-bands_.update(eniric.config.custom_bands)
+bands_.update(config.custom_bands)
 
 
 def band_selector(wav: ndarray, flux: ndarray, band: str) -> Tuple[ndarray, ndarray]:
@@ -306,7 +307,7 @@ def load_aces_spectrum(
 
     Spectra available from http://phoenix.astro.physik.uni-goettingen.de
     """
-    base = eniric.config.paths["phoenix_raw"] + os.sep
+    base = join(config.pathdir, config.paths["phoenix_raw"], "")
 
     if len(params) == 3:  # Only 3 parameters given
         params = [params[0], params[1], params[2], 0]  # Set alpha=0
@@ -320,7 +321,7 @@ def load_aces_spectrum(
         raise ValueError("Number of parameters is incorrect")
 
     wav = phoenix_grid.wl
-    flux, __ = phoenix_grid.load_flux(params, norm=False)
+    flux = phoenix_grid.load_flux(params, norm=False)
 
     # Convert wavelength Angstrom to micron
     wav_micron = wav * 10 ** -4
@@ -390,13 +391,13 @@ def load_btsettl_spectrum(
         assert params[-1] == 0  # Checks index 3 when present.
         params = params[0:2]  # Only allow 2 params
 
-    base = eniric.config.paths["btsettl_raw"] + os.sep
+    base = join(config.pathdir, config.paths["btsettl_raw"], "")
 
     btsettl_grid = BTSETTL(base=base, air=air, wl_range=wl_range)
 
     wav = btsettl_grid.wl
     # CIFIST flux is  W/m**2/um
-    flux, __ = btsettl_grid.load_flux(params, norm=False)
+    flux = btsettl_grid.load_flux(params, norm=False)
 
     # [::10] down samples only every 10th point from BT-Settl CIFIST spectrum.
     wav, flux = wav[::10], flux[::10]