+++ CHANGELOG.md
@@ -26,7 +26,9 @@ Other:
 - Refactoring for 1.0 release.
 - Relax pining in `requirements.txt`.
 - Use `requirements.txt` in `setup.py`
-- move download scripts in scripts/download.
+- Depreciate `--model phoenix` from `phoenix_precision.py`.
+- move download scripts to scripts/download.
+
 
 ### v0.5.1
 ---------